requirements.txt
python-telegram-bot==20.0
requests==2.28.1
watchdog==3.0.0
keyring==23.0.1
keyrings.alt==4.1.1